--- a/extension/parquet/include/column_writer.hpp
+++ b/extension/parquet/include/column_writer.hpp
@@ -27,6 +27,7 @@ public:
 	unsafe_vector<uint16_t> definition_levels;
 	unsafe_vector<uint16_t> repetition_levels;
 	vector<bool> is_empty;
+	idx_t parent_null_count = 0;
 	idx_t null_count = 0;
 
 public: