--- a/extension/parquet/include/column_writer.hpp
+++ b/extension/parquet/include/column_writer.hpp
@@ -113,7 +113,7 @@ protected:
 	void HandleRepeatLevels(ColumnWriterState &state_p, ColumnWriterState *parent, idx_t count, idx_t max_repeat) const;
 
 	void CompressPage(MemoryStream &temp_writer, size_t &compressed_size, data_ptr_t &compressed_data,
-	                  unique_ptr<data_t[]> &compressed_buf);
+	                  AllocatedData &compressed_buf);
 };
 
 } // namespace duckdb