… benchmark/parquet/clickbench_write.benchmark
… extension/parquet/parquet_extension.cpp
… extension/parquet/parquet_writer.cpp
… extension/parquet/writer/boolean_column_writer.cpp
… extension/parquet/writer/enum_column_writer.cpp
… extension/parquet/writer/primitive_column_writer.cpp
… src/common/serializer/memory_stream.cpp
… src/common/types/column/column_data_collection.cpp
… src/common/types/hash.cpp
… src/execution/operator/persistent/physical_batch_copy_to_file.cpp
… src/include/duckdb/common/primitive_dictionary.hpp
… src/include/duckdb/common/serializer/memory_stream.hpp
… src/include/duckdb/common/types/datetime.hpp
… src/include/duckdb/common/types/hash.hpp
… test/api/adbc/test_adbc.cpp
… test/sql/copy/parquet/bloom_filters.test
… test/sql/copy/parquet/writer/parquet_write_memory_usage.test
… test/sql/function/generic/hash_func.test