1. Slightly improve the performance of hashing long strings
2. Implement bit-packing for our `RleBpEncoder` (it only did RLE until now)
3. Implement a custom hash map to build Parquet dictionaries (instead of using `std::unordered_map`)
4. Add many fast paths, e.g., for handling `NULL` values when we know there are no `NULL`s

This brings the time it takes to write `lineitem` from TPC-H SF10 to Parquet from ~4.5s to ~2.5s on my laptop, as well as reducing the size from 2.4 GB to 2.0 GB (with the default `PARQUET_VERSION V1`).

When enabling `PARQUET_VERSION V2`, time reduces from ~3.8s to ~2.4s on my laptop, and size reduces from 1.7 GB to 1.6 GB.

It also adds the parameter `string_dictionary_page_size_limit` to the Parquet writer, which makes the page size of dictionary pages configurable. It defaults to 1 MB.