@gaborcsardi 2025-02-14
Hah, I just finished implementing a better RLE/BP encoder as well, and was about to submit a PR. 😆 
For now it is here: https://github.com/gaborcsardi/duckdb-r/pull/1/files